+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 * Add ISSN, IAN, and other identifier support. Refs [UILD-608].
 * Add ability to change profiles from Instance Edit view. Refs [UILD-576].
 * Add support for publication frequency. Refs [UILD-618].
+* Add read-only editor field support. Refs [UILD-630].
 
 [UILD-552]:https://folio-org.atlassian.net/browse/UILD-552
 [UILD-544]:https://folio-org.atlassian.net/browse/UILD-544
@@ -65,6 +66,7 @@
 [UILD-608]:https://folio-org.atlassian.net/browse/UILD-608
 [UILD-576]:https://folio-org.atlassian.net/browse/UILD-576
 [UILD-618]:https://folio-org.atlassian.net/browse/UILD-618
+[UILD-630]:https://folio-org.atlassian.net/browse/UILD-630
 
 ## 1.0.5 (2025-04-30)
 * Fixed incorrect behavior when navigating between duplicated resources. Fixes [UILD-553].